agent/tiktok_hunter.py
"""
TikTok Hunter - 视频抓取与 AI 视觉分析模块
负责从 TikTok 链接提取产品信息
"""

import os
import re
import asyncio
import tempfile
from typing import Optional, Dict, Any, List
from pathlib import Path

import cv2
import numpy as np
from openai import AsyncOpenAI
from dotenv import load_dotenv

load_dotenv()

# OpenAI Client
openai_client = AsyncOpenAI(api_key=os.getenv("OPENAI_API_KEY"))


class TikTokHunter:
    """TikTok 产品识别引擎"""
    
    def __init__(self):
        self.temp_dir = tempfile.gettempdir()
    
    async def analyze_video_url(self, url: str) -> Dict[str, Any]:
        """
        完整的视频分析流程
        
        Returns:
            {
                "success": bool,
                "product_name": str,
                "category": str,
                "features": List[str],
                "materials": List[str],
                "estimated_price_range": str,
                "sourcing_difficulty": str,
                "confidence": float,
                "raw_analysis": str
            }
        """
        try:
            # 1. 提取关键帧
            frames = await self._extract_key_frames(url)
            
            if not frames:
                return {"success": False, "error": "无法提取视频帧"}
            
            # 2. GPT-4o 视觉分析
            analysis = await self._analyze_frames_with_gpt4o(frames)
            
            return {
                "success": True,
                **analysis
            }
            
        except Exception as e:
            print(f"❌ 分析失败: {str(e)}")
            return {"success": False, "error": str(e)}
    
    async def _extract_key_frames(self, url: str, num_frames: int = 3) -> List[str]:
        """
        从视频中提取关键帧
        
        实际实现中使用 yt-dlp 下载 + OpenCV 提取
        这里提供模拟实现
        """
        # TODO: 实际实现
        # 1. 使用 yt-dlp 下载视频
        # 2. 使用 OpenCV 提取关键帧
        # 3. 保存为临时文件并返回路径列表
        
        # 模拟实现 - 返回占位图片路径
        print(f"📹 正在下载视频: {url}")
        await asyncio.sleep(1)  # 模拟下载时间
        
        # 在实际实现中，这里会返回真实的帧图片路径
        return ["frame_placeholder"]
    
    async def _analyze_frames_with_gpt4o(self, frame_paths: List[str]) -> Dict[str, Any]:
        """
        使用 GPT-4o Vision 分析视频帧
        
        Prompt Engineering 是关键！
        """
        
        # 构建分析 Prompt
        system_prompt = """你是一位资深的跨境电商产品分析专家，专门从视频内容中识别可采购的产品。

你的任务是：
1. 识别视频中展示的主要产品
2. 分析产品的材质、功能、尺寸等特征
3. 评估产品的市场潜力和采购难度
4. 提供专业的采购建议

请用中文回复，并按照以下 JSON 格式输出：
{
    "product_name": "产品名称（中英文）",
    "category": "产品类目",
    "features": ["功能特点1", "功能特点2", ...],
    "materials": ["主要材质1", "材质2", ...],
    "estimated_dimensions": "大致尺寸",
    "target_audience": "目标用户群",
    "selling_points": ["卖点1", "卖点2", ...],
    "estimated_price_range": "估计FOB价格区间",
    "sourcing_difficulty": "low/medium/high",
    "sourcing_advice": "采购建议",
    "confidence": 0.85
}"""

        user_prompt = """请分析这个TikTok视频中展示的产品。
        
视频链接已处理，关键帧已提取。请基于你的专业知识，假设这是一个典型的 TikTok 热门产品视频（如家居小工具、创意电子产品等），进行详细分析。

请特别关注：
- 产品的核心功能和使用场景
- 材质和做工质量预估
- 在跨境电商平台上的竞争力
- 适合的工厂类型（注塑、五金、电子等）"""

        try:
            # 调用 GPT-4o
            # 实际实现中，这里会将图片作为 base64 传入
            response = await openai_client.chat.completions.create(
                model="gpt-4o",
                messages=[
                    {"role": "system", "content": system_prompt},
                    {"role": "user", "content": user_prompt}
                ],
                temperature=0.7,
                max_tokens=1000
            )
            
            content = response.choices[0].message.content
            
            # 解析 JSON 响应
            import json
            
            # 尝试提取 JSON
            json_match = re.search(r'\{[\s\S]*\}', content)
            if json_match:
                analysis = json.loads(json_match.group())
                analysis["raw_analysis"] = content
                return analysis
            else:
                # 返回原始文本
                return {
                    "product_name": "未能识别",
                    "raw_analysis": content,
                    "confidence": 0.3
                }
                
        except Exception as e:
            print(f"❌ GPT-4o 分析失败: {str(e)}")
            # 返回模拟数据用于演示
            return self._get_demo_analysis()
    
    def _get_demo_analysis(self) -> Dict[str, Any]:
        """返回演示用的模拟分析结果"""
        return {
            "product_name": "Anti-Gravity Water Drop Humidifier / 反重力水滴加湿器",
            "category": "Home Appliances / 家居电器",
            "features": [
                "反重力水滴视觉效果",
                "USB 供电",
                "超静音设计 (<30dB)",
                "自动断电保护",
                "7 色 LED 氛围灯"
            ],
            "materials": [
                "ABS 工程塑料外壳",
                "食品级 PP 水箱",
                "超声波雾化片"
            ],
            "estimated_dimensions": "约 15cm x 10cm x 10cm",
            "target_audience": "年轻白领、精致生活追求者、礼品市场",
            "selling_points": [
                "视觉效果独特，适合社交媒体传播",
                "操作简单，即插即用",
                "多功能（加湿+氛围灯）"
            ],
            "estimated_price_range": "$3.50 - $6.00 FOB",
            "sourcing_difficulty": "low",
            "sourcing_advice": "建议选择宁波或中山的小家电工厂，MOQ 通常在 500-1000 件。",
            "confidence": 0.92,
            "raw_analysis": "Demo mode - 实际部署时将使用 GPT-4o Vision 进行真实分析"
        }


class PriceCalculator:
    """FOB 价格估算器"""
    
    # 基础成本系数（按品类）
    CATEGORY_COEFFICIENTS = {
        "Home Appliances": {"material": 1.2, "labor": 0.8, "overhead": 0.3},
        "Electronics": {"material": 1.5, "labor": 1.0, "overhead": 0.4},
        "Textiles": {"material": 0.8, "labor": 1.2, "overhead": 0.2},
        "Plastic Products": {"material": 0.6, "labor": 0.5, "overhead": 0.2},
        "default": {"material": 1.0, "labor": 0.8, "overhead": 0.3}
    }
    
    @classmethod
    def estimate_fob_price(
        cls, 
        category: str, 
        complexity: str = "medium",  # low, medium, high
        quantity: int = 1000
    ) -> Dict[str, Any]:
        """
        基于品类和复杂度估算 FOB 价格
        
        这是一个简化模型，实际项目中应该：
        1. 建立更详细的成本模型
        2. 接入工厂实时报价 API
        3. 考虑原材料价格波动
        """
        coef = cls.CATEGORY_COEFFICIENTS.get(category, cls.CATEGORY_COEFFICIENTS["default"])
        
        # 基础价格（美元）
        base_prices = {"low": 2.0, "medium": 4.0, "high": 8.0}
        base = base_prices.get(complexity, 4.0)
        
        # 计算各项成本
        material_cost = base * coef["material"]
        labor_cost = base * coef["labor"]
        overhead_cost = base * coef["overhead"]
        
        # 批量折扣
        if quantity >= 5000:
            discount = 0.85
        elif quantity >= 2000:
            discount = 0.92
        elif quantity >= 1000:
            discount = 0.96
        else:
            discount = 1.0
        
        total = (material_cost + labor_cost + overhead_cost) * discount
        
        return {
            "unit_price_usd": round(total, 2),
            "breakdown": {
                "material": round(material_cost, 2),
                "labor": round(labor_cost, 2),
                "overhead": round(overhead_cost, 2)
            },
            "quantity": quantity,
            "discount_applied": f"{int((1-discount)*100)}%",
            "lead_time_days": 15 if quantity < 1000 else 25
        }


# 导出便捷函数
async def analyze_tiktok_video(url: str) -> Dict[str, Any]:
    """分析 TikTok 视频的便捷入口"""
    hunter = TikTokHunter()
    return await hunter.analyze_video_url(url)


def calculate_price(category: str, complexity: str = "medium", quantity: int = 1000) -> Dict[str, Any]:
    """计算价格的便捷入口"""
    return PriceCalculator.estimate_fob_price(category, complexity, quantity)
